# backend/swagger.yaml
openapi: 3.0.0
info:
  version: "0.0.1"
  title: Tag Matching App
# during dev, should point to your local machine
servers:
  - url: http://localhost:3000/api/v1/
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
security:
  - bearerAuth: []
paths:
  /users/register:
    post:
      description: Register a user account
      operationId: registerUser
      tags:
        - authentication
      security: []
      requestBody:
        description: The user object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: The user ID
                  emailAddress:
                    type: string
                    description: The email address
                  password:
                    type: string
                    description: The hashed password
                  __v:
                    type: string
                    description: Some user data
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
  /users/login:
    post:
      description: Log in with a user account to get an access token.
      operationId: loginUser
      tags:
        - authentication
      security: []
      requestBody:
        description: The user object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The JWT access token
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
  /users/me:
    get:
      description: Get the user account.
      operationId: getUser
      tags:
        - authentication
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The user ID
                  emailAddress:
                    type: string
                    descrition: The user's email address 
  /like/{userId}:
    post:
      description: Add a like from one profile to another
      operationId: addLike
      parameters:
        - name: userId
          in: path
          description: User ID to like
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Success
          schema:
            type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
    delete:
      description: Unmatch a profile
      operationId: deleteLike
      parameters:
        - name: userId
          in: path
          description: User ID to unlike
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
  /find:
    get:
      description: Returns a set of profiles fitting criteria for a profile id
      operationId: findProfiles
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ProfileObject"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
  /profile/{userId}:
    get:
      description: Returns a profile by id
      operationId: getProfile
      parameters:
        - name: userId
          in: path
          description: User ID of the profile to retrieve
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/components/schemas/ProfileObject"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
  /profile:
    post:
      description: Create a profile
      operationId: createProfile
      requestBody:
        description: The user object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileObject"
      responses:
        "201":
          description: Success
          schema:
            type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Error
          schema:
            $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      properties:
        emailAddress:
          type: string
        password:
          type: string
    ProfileObject:
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
          description: YYYY-MM-DD To be formatted by the frontend
        occupation:
          type: string
        bio:
          type: string
        photos:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        maxDistance:
          type: number
          description: In kilometers.
        ageRange:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
  responses:
    ErrorResponse:
      properties:
        statusCode:
          type: integer
        errorCode:
          type: integer
        errorMsg:
          type: string
        raw:
          type: string
    UnauthorizedError:
      description: Access token is missing or invalid
  